requirements.txt

# Core ML Dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0

# Data Processing & Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pandas>=1.3.0
opencv-python>=4.5.0

# Web Interface
streamlit>=1.25.0
plotly>=5.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Model Export (Optional)
onnx>=1.12.0
onnxruntime>=1.12.0

# System Utilities
psutil>=5.8.0
python-dotenv>=0.19.0
pathlib2>=2.3.6

# Development Tools (Optional)
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7.0
flake8>=3.9.0
mypy>=0.910
types-Pillow>=8.3.0

# Jupyter Support (Optional)
jupyter>=1.0.0
ipykernel>=6.4.0
ipywidgets>=7.6.0

# Additional Utilities
requests>=2.26.0
pyyaml>=5.4.0
tensorboard>=2.7.0
